Release 0.1.8
This is mostly a Linter/Docs/Comments fix release.
There is only one code change: remove the custom validator that was enforcing the None -> [] rule for both tags/tag_ids.
It turns out that it's actually easier to keep the distinct meanings of `[]` and `None` and work around the inconsistencies elsewhere.

This also does a small dep bump.